router: change URL in between navigating & navigated (#4974)
With this change, we now modify the URL before `navigated` is fired so
all the downstream can send request right away without having to confirm
that the route change is indeed reflected in the URL bar.

Context: previously, our router modified the URL only when the
`navigated ` action is fired. Often, this action is listened to by many
views/effects so they can do a cleanup or a data fetch for the new
route. Now, because the URL was only modified after `navigated ` is
fired, there could be a timing issues where, depending on the order of
execution, we can be making requests to the unmodified URL, causing
issues.

Now, we are changing the semantics of the `navigated` to really really
mean, 'hey, the navigation is now complete and everything is settled'.

Tested this change internally.