timeseries: add data source for persisting setting
Background
----------
Currently, TensorBoard forgets settings (at least on the Angular side)
when browser refreshes and even for the ones that are persisted in the
URLs do not carry over to other experiments or other TensorBoard browser
tabs.

This change
-----------
This change is a part of the greater effort to mitigate the usability
issue by persisted few settings (for now) to LocalStorage. Do note that
this is not a panacea as users can change port number and the states
will be forgotten.

The DataSource API is created with hosted services in mind where we may
add a HTTP API or DataProvider API for persisting user settings.