timeseries: persist and read settings
This is a continuation of #4991 where we are attempting to persist
several settings from TimeSeries to improve frustrating user experience
of having to reset settings over and over.

This change integrates the DataSource with the MetricsEffects so persist
new settings when it changes and bootstrap the application state from
the state in the LocalStorage.

Product choice I took that may be up for discussion is, this change
persist the smoothing value user may provide in the URL since it felt
more natural that way.