timeseries: improved formatting on tooltip
Previously, we were showing close to the "raw" value in the tooltip
often leading to very long nonsense especially with the floating point
math errors. The impact of said errors are often seeing a number like
'0.3000000000004' where the last digit `4` often is meaningless. To work
around this, we are using d3.format to remove unnecessary precision
which just often adds visual noises.