sync: fix object property accessor (#5007)
`unsanitizedObject` was typed as a Record or just an object whose
property we are trying to access. To make it survive the mangling, we
need to supply proper typing information for the `JSON.parse`d object
which is now assigned as `SerializableSettings` which is capable of
surviving the mangling (`declare`d).