timeseries: fix runs table scrollable issue
When a `<runs-table-component>` has height = 100px, the .table-container
also had height = 100px which is not what we intended. We intended it to
have 100px - heightOfFilterRow. We now use the proper flex syntax for
the proper layout.

This purely visual bug is hard to test in unit test and will be covered
via screenshot tests internally.