runs: create groupBy override vs. initial value (#5018)
When we work on our setting persistence, we want to be able to change
our default behavior. However, if we always persist the groupBy value in
the store to the URL, current business logic will be ingrained into
every dashboard URLs and its bookmark can lead to a case where we cannot
change our default behavior.

this change simply refactors so initial value of `groupBy` is separated
from the overridden value.