dark mode: fix hparams, graph, mesh, and profile
This change fixes the visuals of hparams, graph, mesh and profile
redirection plugin under the dark mode.

Do note that we did not attempt to have a generic solution for all
iframe based plugin since it becomes wildly more difficult to reliably
apply dark styles. Instead, we will show the default light mode which
will be quite jarring but that's the best we can do as of right now.