metrics: separate default settings from override
Background: currently, from view's and deep linker's perspective, there
is no way to discern a setting that was overridden vs. default. This
makes the deep linking particularly finicky since, if a different
application provides a different default for smoothing, for example, it
will show up in the deep link as it does not match OSS's default 0.6.

Change: this change separates default away from overrides while making
it largely transparent to views.