debugger: fix mutating object by ref in reducers
In debugger v2 reducers, we were, by mistake, modifying values of
previous state. This mistake manifested in a form of runtime error for
writing to the read-only property.

This mistake was corrected by testing with a frozen objects in the
tests. Should we make more mistakes like this in the future, we may
really want to consider requring all reducer tests to use `deepFreeze`.