dark mode; allow dark theme to use different vars
Theme objects are created from `tb` variables like `$tb-primary`,
`$tb-accent` and others. These variables, however, may not look the best
again the dark background and we have to support different values for
dark version.

This change makes sure that the variable like `$tb-dark-primary` are
defined if it is not already defined to their light counter part,
`$tb-primary`.

This change applies the change by making the primary color a bit darker
orange, making it look better against the dark theme.