debugger: deflake debugger v2 test (#5053)
Now that the debugger v2 depend on the feature flag feature (dark mode
in the SourceFileContainer), we have to provide appropriate stub or
remove it in the `import` path. This change refactors how the test is
authored so we do not have to stub out selectors from other containers.

Source of the error
===================

The error of 'accessing property `defaultFlags` of
undefined' was happening reliably but was throwing in the `afterEach`
which has made the test runtime/reporter behavior flaky (when we were
lucky, the test failed and when we were unlucky, the exception slipped
through the crease). We were able to get the error reliably when run
with `runs_per_test=100` though.

More importantly, the source of the error is the structure of the test.
The unit test had imported the real sub modules and has loaded the
real source-file-container which DebuggerComponent happens to use. This
has made the real implementation of the source-file-container access the
real data without appropriate stub implementation provided via
`overrideSelector`. To fix this, we had made sure no extraneous modules
are loaded as part of the test and only focus on the implementation
current module provides.