source-code: bug fix where dark mode did not apply
Previously, when manually tested against the only user of the feature,
the source-code component lazily created the monaco editor when the data
was present. However, as I have refactored the code a little bit, the
assumption has toppled and resulted in a bug where initial values are
not set properly. The order of events are following:

- source-code component gets mounted and `ngOnChanges` get triggered
  with the theme value. Because the monaco is not fetched yet by the
  container, it early returns and theme value never gets set.
- when source-code container finally loads the monaco editor,
  ngOnChanges only include the value, `monaco`, and our theme applier
  never gets triggered.

Fix this by detecting the condition in which we have to apply the
settings to the editor.