theme: allow customizable `app-bar` theme var
This is the final iteration of the change. Previously, we had thought
that using `default` value to `mat-color` would be sufficient but we
were incorrect.

Background is the following. In other variants of TensorBoard, we
elected to use very dark primary color. It looks nice under the light
theme but under the dark theme, its lightness is too close to that of
background so we had to make its primary default color to be light. With
previous change, now, our app-bar under dark mode would be super light
which is certainly not what we had wished.

With other hackier theme template munging set aside, now, we are
choosing to declare `app-bar` as one of special variable that can be
customized like `tb-primary` that is, when provided, will be used in the
theme object.