core: use router event to fetch data
This reverts commit 06c22d6fa7a1a5512d45e8015d6174d2e2020b1a.
Related GitHub pulls are: #4915, #4935, #4941, and #4953.

In the first batch of rollback, we had broken one app and, in the second
batch, we have broken another TensorBoard app. Now, we have sufficient
acceptance test to have a confidence in this approach.

Description from the original change:

```
Context: previously, we were using a composite action[1] to know when
the dashboard view updated and thus its data needs to be refreshed.

Fix: now that TensorBoard has a router and can determine what views are
rendered given a route, we can use a purer redux pattern to determine
whether the dashboard data has to be reloaded. Since not all versions of
the app uses the router, we will temporarily support the coreLoaded.

[1]: composite actioness: view was subscribing to the state changes to
determine to fire an action which in turn triggers another action.
```

TODO:
- [x]: test sync: cl/379536638
- [x]: manual acceptance test on one project