dark mode: enable media query based dark mode (#5074)
This change enables the dark mode in OSS. The feature will turn on by
default when user's browser says they prefer the dark mode.

The feature can be mnaually turned off with `?darkMode=false`. In the
future, we will add a button that lets user toggle the dark mode in the
header.