route: omit displayName from routeId.
RouteId is a unique opaque identifier for a route. Previously, we
included display names as part of the unique identifier but we decided
to drop it for improve the UX.

Specifically, we would like to be able to let user change display name
dynamically without causing a routeId changes (which will cause router
to actually do a route change).