Move settings redux to settings feature. (#5090)
There are several properties that are global user settings that are currently located in the "core" feature:
* reloadEnabled
* reloadPeriodInMs
* pageSize

We move these and all associated redux to the "settings" feature. We also migrate some references to them elsewhere in the code.

We get a bit lucky with google-internal code. This change is written so that we don't have to modify any google-internal code immediately. We:
* Reexport some settings actions and selectors from core.
* Hack CoreState type so that its declaration includes SettingsState.

Everything seems to build and work fine with these changes. (Googlers, see cl/381226386). After changes here are imported into Google, we will update internal references and then another PR will remove some of the hacks introduced in this change.

Testing:
* Migrated all relevant tests from core to settings.
* Ran tensorboard and played with settings and see that they work.
* Ran internal tensorboards (tbdev and tbcorp), see that they work, and see that their analytics code also still works.