core: set correct initialState in reducers
Depending on order of definition, when using routeContextedReducerUtil,
there are cases when the initialState is completely empty. This happens
because the first instance of the initialState in the `composeReducers`
get used.

While it might be advisable not to specify and rely on order of reducers
defined in the `composeReducers`, the order is not specified as part of
the spec and TensorBoard specifies them in arbitrary order. In such
case, we don't want a broken or misbehaving application with missing
initial value.

As such, we now populate the correct initialState and wrote appropriate
test that would have caught the issue.