router: allow onRouteIdChange to reason with route
Previously, the optional third parameter to the route contexted reducer
helper only allowed the utility to specify a state [re]setter that
reasons with the state alone. However, there are cases when we need to
set the initial value of a store based on the routeKind. For instance,
we would like our colorGroup settings to be `GroupBy.EXPERIMENT` when
user is navigating to an `COMPARE_EXPERIMENT` routeKind while keeping
it `GroupBy.RUN` in other cases. Since the initial state is a function
of the new route (`routeKind`, really), this change expands the
API signature to allow for reasoning with the route.