line-chart: set attribute when data is loaded once
For better WebDriver testing, we would like to distinguish not_loaded,
loading, and loaded_at_least_once. By using a separate atttribute on the
vz-line-chart, we were able to discern these differing states and
utlimately made our tests a bit more stable.

This change introduces another internal state in the
data-loader-behavior with which we can set the attribute onto the
vz-line-chart.