persistent_settings: glue config and data_source with Redux (#5134)
This change includes Redux changes that glues the Settings configuration
with DataSource. When there are configurations for global settings, the effects
now subscribe to the store changes, and when value changes, we persist the
setting in the local storage using the DataSource.

Note that this change does not yet migrate Metrics or notification_center to
use the new persistent_settings.