router: redirect with query parameter
This change adds advanced feature to the router where a redirector can
redirect to a path with query paramter.

Imagine following usecase. We want to add a nice redirection route
where, depending on path part value, we can redirect to EXPERIMENT or
EXPERIMENTS and, in the latter case, from the path part value, we'd like
to also prepopulate the user token. This would require
DeepLinkProvider-like module that would also reason and prepopulate
state from a path part (as opposed to value in the query parameter) and
that's farther in the realm of possible designs than what is proposed
here.

Here, we are proposing that the redirector can create almost
`Navigation` like payload where it can specify both pathname and query
parameter to a next route. Do note that this feature is clamped down to
only programmatical navigation and we have no plans to expand scope of
other route definitions.