Expose CompareExperimentStr parser as util
Ideally, we want to populate store with an information that exists in
the comparison string. However, router only provides the information in
an opaque `experimentIds` (typed as string) without an ability to reason
with the structure inside.

Ideally, RouteParameter should have a way to construct a structured data
structure and pass it through `navigated` action but that makes
AppRoutingEffects even more complex while making the RouteConfig more
elaborate.

For now, we are exposing the utility for stores to use instead.