histogram: axis for Angular histogram widget (#5159)
This change is a start of the rewrite for the histogram widget.
Currently, the component was written as a shallow wrapper around our
Polymer based vz-histogram-timeseries component. This wrapper was not
easy enough to make changes to address some of the user feedbacks and
implement a new feature.

To that end, we are rewriting the component so bulk of the logic resides
in the Angular land without Polymer. Do note that goal is _not_ to
remove Polymer but is to have more manageable visualization code with
much better types. Even the Polymer component is a thin wrapper around
d3 and the most interesting logic resides in the `draw method`. To be
clear, we are not completely gutting out d3 as it still serves useful
purposes for axis, formatting,

This change is the first of more to follow. This change only calculates
the axis and renders the same axis as existing histogram visualization.
No component actually make use of it and you will need to manually
modify `histogram_card_component.ng.html` to make use of the new
version.