histogram: render Overlay and Offset charts (#5162)
This change builds on top of the Angular version of the histograms
component where we had axes rendering implemented. This change finally
draws the histogram like the Polymer counterpart. Most of the UX
considerations are mostly replica of that of Polymer part and I did not
intend to have much changes besides minor color changes (now use theme
variable instead of hardcoded hex value).

This change does not include:
- interaction like tooltips
- redraw on resize