histogram: prevent render while invisible.
This change integrates Intersection Observer to the new Angular based
histogram so we prevent needless updates while invisible. Do note that
this solution is not perfect because any prop changes will be reflected
onto the DOM created via template; could not yet find an Angular
equivalent to `shouldComponentUpdate` in React.