color group: set run color for ones without color assignment
With the color groupign feature, not all runs have color assigned. For
those without a color assignment, we would like to show gray or white
color and it is now a responsibility of the view to choose the color.
This change makes it so that when a color is not present in the
assignment map, we return `#fff` instead of throwing an invariant error.