linked time: range on sidebar step selector
This change introduces a new state that keeps track of minimum and
maximum step values from all time series data. This is so we can render
the range_input with its minimum and maximum according to the data.

This change also updates the data structure on the `LinkedTime` to drop
`wallTime` field which is not used meaningfully and only causes the data
inconsistency.