linked time: marker when extent is outside of linked time
We now show a small read "i" when an x-axis extent (extremums) of the
line chart is not intersecting with the selected time. This "i"
informs user that the data presented below is not within the linked
time. In that situation, we will clip the selectedTime to the minimum or
maximum of the dataset so something is visible on the screen.