line chart: zoom instruction show on top of other (#5215)
Especially with the customizable interactive view, the line chart
interactive view is no longer on top of all the other DOM. This makes
our zoom instruction which should appear above all is getting clipped or
getting masked by other components. This change simply uses z-index: 1
(we do not use it elsewhere so 1 is sufficient; we rely on DOM ordering
for the most part) to lift the DOM up.