timeseries: filter card by visualization
This change implements UI for filering the TimeSeries visualization by
type of the plguin. This is so that we can later deprecate dedicated
plugins for a specific plugin type while having comparable experience
with the TimeSeries plugin.

Do note that, for now, I decided not to apply the filter on the pinned
visualization as it is yet unclear whether the pinned card should also
be filtered.

The UI may not be the most ideal and we may want to iterate and hone its
placement and UX.