timeseries: filter card by visualization (#5222)
This change implements UI for filtering the TimeSeries visualization by
type of the plugin. This is so that we can later deprecate dedicated
plugins for a specific plugin type while having comparable experience
with the TimeSeries plugin.

Do note that, for now, I decided not to apply the filter on the pinned
visualization as it is yet unclear whether the pinned card should also
be filtered.