chore: require node.js 12.21.0 (#5231)
By default, no matter what version of Node you have installed,
rules_nodejs loads node 12.13.0. This version is not too compatible with
the newer dependencies (a package may use `engine` on package.json to
denote compatibility with a version of a node runtime). This change
bumps the node requirement to 12.21.0 which is good enough for our use
case.