timeseries: improve card observer visibility logic with element ids (#5234)
The Time Series dashboard has been under the assumption that `getVisibleCardIds`
accurately returns the set of CardIds that are within the visible viewport.

This assumption, backed by the CardObserver using an IntersectionObserver, is
false in an event sequence involving the same CardId twice:
- domElementA with cardId1 is shown
- domElementB with cardId1 is hidden

The Ngrx store is misled to believe that cardId1 is hidden, even though
domElementA may still be visible.  This change corrects the CardObserver to
track unique DOM Elements.

The symptom is flaky and hard to consistently reproduce.  Manually tested by
patching in https://github.com/psybuzz/tensorboard/tree/persist-tag for tagFilter
persistence in the URL, opened URLs with "?tagFilter=accuracy" that cause the
bad behavior more consistently, and observed that they no longer occur.

Googlers, see b/179450711.