notification: clip the date to today
There were cases when the notification payload we handcrafted is in the
future and since we only dismiss the notification by recording the last
read time (NOW), there were cases when new notifications in the future
are never dismissable. To prevent developer mistakes, we can clip the
date to start of the day so a notification can never been in the future.