time series: deduplicate tags in the search (#5249)
Image cards are a bit odd; unlike others, the same tuple of run and tag
can appear more than once and that is because it has the third
axis--sample. When there are more than one sample for a run+tag, the tag
filters previously showed the duplicate entry which this can
de-duplicates.