timeseries: run selection based on regex filter (#5252)
Previously, regex filter was mainly to influence the list you see on the
runs table. This is where it differs from the traditional runs-selector
(Polymer) which not only filtered runs but also changed the runs
selection that you see on the chart. This change brings the Polymer
behavior to Angular as users have preferred that.