timeseries: show pinned view while filtering (#5255)
You may want to treat the pinned UI as the shopping cart and want to add
cards to the "cart" while filtering. In the past, the shopping cart did
not show up when filtering so it was difficult to understand all the
items in the shopping cart. To fix that, we now show pinned cards below
the filter view each with their own scrollable container.