histogram: use `transform` to prevent repaint
Paint is one of the most expensive operations when you have too many
items painting at the same time. One technique to prevent that is to use
`transform` and `will-change` which often put the DOM in the GPU context
instead. All in all, using `attr.transform` still causes repaint while
`style.transform` does not so we chose to use latter instead.