timeseries: improve first paint
This change adopts the "isEverVisible"-ness to all types of
visualization. The premise of this change is following: some users wants
high number of pagination size with many distinct tag groups. For those
users, we optimized line chart render by deferring render until the card
is visible on the screen. With the optimization, we were able to reduce
number of DOM nodes we mount at the start of the page load. This change
expands the optimizations to defer rendering card headers and rendering
other visualization types like image and histograms.