timeseries: improve tag filtering
In performance profiling, there were two medium sized opportunities for
optimization:
1. we were spending significant CPU cycles for rendering all tag matches
in the autocomplete and it was taking non-trivial amount of time
(~100ms) when there were 3000 tags. This definitely contributed to
slowness when every keystroke is expected to be very responsive.
2. `compareTagNames` were taking a significant amount of time. Given
that the list of tags/cards does not change unless the underlying data
from backend changes, we really do not need to sort on every keystroke
that changes tagFilter. We can pre-sort the tag list before we apply the
filter.

These two changes, combined, gave significantly subjectively snappier
experience.