timeseries: debounce tag filter changes (#5265)
This puts debounce on the tag filter changes. Tag view does not have to
change on every keystroke so we put some debounce period so we can be
more performant by not updating too much.

This change also moves the plugin type filtering up since it does not change
too rapidly.