timeseries: show spinner when tags load (#5271)
When the initial tags are loading and when it is taking a long time,
users do not see anything for awhile. This can lead to a confusion and
lead people to believe that something in the plugin rendered
incorrectly.

Now, for the initial load only, this change shows spinner for the first
render.