timeseries: implement "go to scalars" (#5295)
This change implements "Go to scalars" by allowing core reducers listen
for the "metricsGoToScalars" action. While it may look odd for a
reducer to subscribe to an action from another feature, this is natural
since actions are just an interface/function and core plugin can
function even without, for example, Metrics feature present in the
application.