infra: add strict_templates checking for Angular
This change adds a line to tsconfig that allows Angular to check types
more strictly in the template. This is relatively new option so it
remains to be seen how capable and power this feature is.

For more information, refer to
<https://angular.io/guide/angular-compiler-options#stricttemplates>

This change includes one instance of a violation that was found in the
internal version of the app.