timeseries: partially implement linked time on histogram (#5299)
This change implements linked time fob on the histograms visualization.
Unlike the scalar card case where we really want to make a generic line
chart component, this change embeds the linked time feature in the
histogram visualization because, that way, we do not have to contort
ourselves with its complex coordinate system. This makes the code a lot
more manageable and easy to understand.

While implementing the feature, this change refactored the visual
component for the linked time fob so we can reuse its style in other
places too.