timeseris: fix tags not fetching issue when loading
Previously, we were only waiting for navigation and plugin change events
none of which is useful when freshly loading dashboard since
`activePlugin` will be null until we can fetch the plugins listing at
least once. This change, to fix the problem, also listens to plugins
listing loaded event to assert and fetch the data.