timeseries: theme group control (#5298)
Pagination controls were improperly themed and this change fixes that.
Now, the secondary-text colors are correctly picked based on current
theme settings and the `<input>` is hard themed (instead of relying on
OS settings and browser's default dark styling). It also styles disabled
button so when the "Previous" or "Next" button is disabled, we have a
visual cue.