linked time: implement histogram color for linked time (#5302)
This change implements fill/color for the linked time feature. When in a
single step mode, if there exists a histogram data with the matching
step, we will highlight it by putting subdued color for non-matching
ones while showing a color for the match. For the range selected time
mode, this change does the same for match/non-match creating a range of
colored histograms.