timeseries: show indication when there is a new pin
The reason why we decided to show pinned section during search is
because we wanted to make pinning more intuitive while you are filter
searching (previously, when you pinned a card, there was no visual
indication of an action besides a small fill on the pin icon which is
occluded by the cursor). This change makes the section a bit smaller but
compensate for the small UI with an indicator for new pinned card.