timeseries: show relative time in tooltip
Currently, we are showing relative time in the tooltip only when you are
using x-axis type == relative. Instead of conditioning like that, we are
now showing the time all the time. This change, to support that,
refactored the data type a little bit to always compute the relative
wall time between the point (which required a lot of test refactors).