dark mode: Style link colors in both dark and light modes in polymer plugins. (#5318)
As reported in #5309, links in some error messages in dark mode are difficult to see. It turns out this is potentially a more general problem: there is no attempt to style any links for dark or light mode in the polymer part of the code base.

Mirroring definitions for the Angular part of the code base we add --tb-link and --tb-link-visited definitions in tensorboard-color.ts.

The corresponding Angular definitions can be found in:
https://cs.opensource.google/tensorflow/tensorboard/+/master:tensorboard/webapp/theme/_tb_theme.template.scss

We use --tb-link and --tb-link-visited in dashboard-style.ts, applying the styling for most dashboards.

We use --tb-link and --tb-link-visited in the graphs and hparams dashboards, since they do not import styles from dashboard-style.ts.