timeseries: Adjust scroll position after pagination (#5301)
When the moving to the final page in a cardGroup there is sometimes a jarring effect caused by the changing of the number of cards. This often will even cause the cardGroup to move completely offscreen which is confusing for users. This change adjusts the scroll position after pagination so the control used to turn the page stays in the same position on the users screen.