runs: remember color by id instead of hex (#5338)
This is feature/visual no-op refactor that addresses #5306. As mentioned
in the issue, we have been remembering hex values for colors which made
dynamic color palettes difficult; we want to use different set of colors
when in dark mode and when user supplies a palette in the future. This
change refactors how run colors are remembered in the runs reducers so
we can later also remember chosen color palette.