color: put ColorPalette as state and improve data structure
This change makes a color palette more elaborate so you can have a light
and dark variant of a color with an id and display name. This is so that
user can use different sets of colors without modifying our source code.

After this change, we will introduce a new setting in the dialog that
let user choose and preview different color palettes.