timeseries: match selection using same logic (#5334)
Previously, to allow regex string also match run selection, we have
extend the utility selector to match run selection against the regex.
The logic, however, was massively flawed--instead of matching against
name of a run, experiment, alias, and legacy run name, the
implementation matched against runId which coincidentally contained run
name but was supposed to be an opaque id. This caused an obvious match
like `/^foo/` not to match anything especially when an opaque id is not
created to start with "foo".

In the end, this change basically re-implements the run selection based
on regex more correct.

In the process, we refactored how a run is matched against regex so the
behavior is more congruent in our app.