run selector: subdue violent exception when hparams are invalid (#5344)
* run selector: subdue violent exception when hparams are invalid

When hparams specs in comparison mode are not compatible (e.g., an
experiment has Discrete one while for the same name another experiment
has Interval one), we currently throw a `RangeError`. That error manifests
as very violent error that breaks entire app and make the TimeSeries
plugin not functional. Since TimeSeries does not really render hparams,
it does not make sense for this error to so violently throw error. Now,
with this change, we do not even read the hparams value when in
TimeSeries.

Also, as a bonus, we have relaxed the constraint a bit and try to
reconcile with incompatible filters if any of the compatible ones are
empty.

* grammar