runs-table: do not match against exp name (#5351)
Runs table, when in comparison mode, is capable of rendering experiment
alias and show the full experiment name in the hover text. However,
because we are matching regex against experiment name that is not
visible without a user interaction, the match is quite surprising and is
surprising users.

This change removes an ability to match against the experiment name.