settings: customizable color palette
TensorBoard currently ships with a set of colors for visualization. One
of the key limitations and consideration was that we wanted the default
experience to be respectful to a person with color blindness. As a
result, our color set only contained seven colors and we have received
many feedbacks that the set is too repetitive.

This change mitigates the issue by allowing users to customize and
select more expanded set of colors. However, this change does not yet
remember the selection and will be reset every time the application
opens. This problem will be solved in a subsequent change.