line chart: fix fixedViewBox flow (#5367)
fixedViewBox is way for parent to set the viewBox so it is not computed
from the data. This is approrpriate for some use cases like PR curve
where the axis should always span [0, 1].

Previously, it was broken when there was a resize on the window or when
data series changed which was not an intended behavior. This change
fixes that.