persistent settings: expose OSSSettingsConverter as DI-able (#5387)
Currently, OSSSettingsCOnverter is only acquirable by injecting a
symbol, SettingsConverter. If we want to compose the converter in a type
safe manner, we have to convert the converter to be a generic or compose
by injecting an instance of the converter. Here, we have decided to take
the latter approach which tends to keep the base implementation simple
to reason with without generics.