layout: make resizer smaller (#5390)
Previously, we used to have a little icon and thick vertical bar to make
resize affordance apparent on the left pane resizer. After surveying
modern apps, it looks like they are often done with pointer in
combination with thicker vertical bar on hover so we adopt those
designs.

In the end, with this change, we give users about 18 pixels back in the
horizontal space.