bug: prevent violent error when Experiment metadata is missing (#5393)
ui_selector, which provides a convenience selector for getting the run
selection, was throwing a JavaScript exception violently when experiment
metadata did not load but runs list was loaded.

Above case can happen when the backend implements /data/experiments
incorrectly and inadvertently apply `limit` on the list of experiments
passed. As such, TensorBoard may not fetch an experiment metadata ever
when list of experiment ids go above `N`.

In theory, there should be a fix to the backend, but we also do not want
the web code to violently throw and henceforth break the app completely.
Especially since the web code never makes use of the experiment name
anymore when filtering, we can easily guard against the case, hence the
change.