timeseries: fix button toggle state issue (#5398)
Angular material components are not truly "controlled input" that even
if the parent sets `checked` state to false, if user clicked on the
button from false to true to cause a redux state change that makes
parent put that state, it will have a value `checked`. This is directly
observable by printing out value of `checked` vs. value set by parent.

We work around this issue by simply using button that acts like a
checkbox instead of using an Angular material component. Also, styled
carefully to make it look about the same.