widget: content wrapping input (#5407)
This change open sources content_wrapping_input, a component that
behaves a lot like content-editable element that tightly wraps content.

A conventional `<input>` element has a fixed width invariable to size of
the content. If content is too long, there exists mechanisms to scroll
within the input. This behavior is not something that we do not want in
some UI elements and this widget helps with that.