[time-namespaced state] Generate namespace ids in app_routing_effects. (#5435)
As part of the work to migrate to "time-namespaced state" (Googlers, see http://go/tb-timenamespaced-state) we need to refactor namespace id generation from route_contexted_reducer_helper.ts to app_routing_effects.ts.

This requires changing the navigated() action to include the namespace information - beforeNamespaceId and afterNamespaceId. So there is a somewhat high blast radius. I've tried to ensure that the changes will not break the internal build.

There is also a new property in app_routing state: activeNamespaceId. We use this to populate the beforeNamespaceId value.